requirements.txt
# annotated-types==0.7.0
# anyio==4.8.0
# bcrypt==4.0.1
# beanie==1.29.0
# certifi==2025.4.26
# click==8.1.8
# dnspython==2.7.0
# exceptiongroup==1.2.2
# fastapi==0.115.11
# h11==0.14.0
# idna==3.10
# lazy-model==0.2.0
# motor==3.7.0
# passlib==1.7.4
# pydantic==2.10.6
# pydantic-settings==2.9.1
# pydantic_core==2.27.2
# PyJWT==2.10.1
# pymongo==4.12.0
# python-dotenv==1.1.0
# python-multipart==0.0.20
# sniffio==1.3.1
# starlette==0.46.1
# toml==0.10.2
# typing-inspection==0.4.0
# typing_extensions==4.12.2
# uvicorn==0.34.0



annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.0.1
beanie==1.29.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lazy-model==0.2.0
motor==3.7.0
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
toml==0.10.2
trio==0.30.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
